Minor changes to add-on features
Added the term "Add-on" to "Features" title to clarify that these are features that could be implemented after MVP is completed, not features that are part of the MVP. Note that Features was already a separate section from the MVP section.

Changed language of first add-on feature to clarify the 'like' / 'dislike' spirit of this feature. We are not using personality types like those proposed by the Myers-Briggs Type Indicator.